Remove drupal trait, merge with init
