+++ src/Command/MakeCommands.php
@@ -82,7 +82,7 @@ class MakeCommands extends Tasks
         foreach ($targetsAndLinks as $targetAndLink) {
             $this->docrootSymlink(
                 $targetAndLink['target'],
-                self::$drupalDocroot . $targetAndLink['link']
+                'docroot' . $targetAndLink['link']
             );
         }
     }
@@ -96,10 +96,10 @@ class MakeCommands extends Tasks
         $link_dirname = $link_parts['dirname'];
         $target_path_relative_to_link = (new Filesystem())->makePathRelative($target, $link_dirname);
 
-        if (!file_exists($target) || !file_exists(self::$drupalDocroot)) {
+        if (!file_exists($target) || !file_exists('docroot')) {
             $this->io()->warning(
                 "Skipping linking $target. Folders $target and '" .
-                self::$drupalDocroot . "' must both be present to create link."
+                'docroot' . "' must both be present to create link."
             );
             return;
         }
